pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "antenna-classifier"
version = "0.1.0"
description = "NEC file parser, validator, and antenna type classifier"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = []

[project.optional-dependencies]
dashboard = [
    "fastapi>=0.110",
    "uvicorn>=0.29",
    "psycopg2-binary>=2.9",
    "openai>=1.30",
    "pdfplumber>=0.11",
]

[project.scripts]
antenna-classifier = "antenna_classifier.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
antenna_classifier = ["static/**"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
